+++ R/plans.R
@@ -101,7 +101,7 @@ get_plan <- function(countries, additional_exemplar_countries = NULL,
                      iea_data_path, ceda_data_folder, exemplar_table_path,
                      fu_analysis_folder, reports_source_folders, reports_dest_folder) {
 
-  # Get around some warnings.
+  # Get around warnings of type "no visible binding for global variable".
   alloc_and_eff_couns <- NULL
   map <- NULL
   AllIEAData <- NULL
@@ -116,6 +116,8 @@ get_plan <- function(countries, additional_exemplar_countries = NULL,
   ExemplarLists <- NULL
   CompletedAllocationTables <- NULL
   CompletedEfficiencyTables <- NULL
+  Cmats <- NULL
+  EtaPhivecs <- NULL
 
   p <- drake::drake_plan(
 
@@ -224,6 +226,19 @@ get_plan <- function(countries, additional_exemplar_countries = NULL,
 
     # (10) Extend to useful stage
 
+    Cmats = drake::target(calc_C_mats(completed_allocation_tables = CompletedAllocationTables,
+                                      countries = countries),
+                          dynamic = map(countries)),
+
+    EtaPhivecs = drake::target(calc_eta_fu_phi_u_vecs(completed_efficiency_tables = CompletedEfficiencyTables,
+                                                      countries = countries),
+                               dynamic = map(countries)),
+
+    PSUT_useful = drake::target(move_to_useful(psut_final = PSUT_final,
+                                               C_mats = Cmats,
+                                               eta_phi_vecs = EtaPhivecs,
+                                               countries = countries),
+                                dynamic = map(countries)),
 
     # (11) Add other methods
 
@@ -253,7 +268,7 @@ get_plan <- function(countries, additional_exemplar_countries = NULL,
     EfficiencyGraphs = drake::target(eta_fu_plots_df(CompletedEfficiencyTables, countries = countries)),
 
     # Build Exergy-to-energy ratio graphs
-    ExergyEnergyGraphs = drake::target(phi_u_plots_df(CompletedEfficiencyTables, countries = countries))
+    # ExergyEnergyGraphs = drake::target(phi_u_plots_df(CompletedEfficiencyTables, countries = countries))