+++ tests/testthat/test-plans.R
@@ -64,10 +64,13 @@ test_that("get_plan works", {
   expect_equal(my_plan[[rn <- rn + 1, "target"]], "IncompleteEfficiencyTables")
   expect_equal(my_plan[[rn <- rn + 1, "target"]], "CompletedEfficiencyTables")
 
+  expect_equal(my_plan[[rn <- rn + 1, "target"]], "Cmats")
+  expect_equal(my_plan[[rn <- rn + 1, "target"]], "EtaPhivecs")
+  expect_equal(my_plan[[rn <- rn + 1, "target"]], "PSUT_useful")
+
   expect_equal(my_plan[[rn <- rn + 1, "target"]], "AllocationGraphs")
   expect_equal(my_plan[[rn <- rn + 1, "target"]], "NonStationaryAllocationGraphs")
   expect_equal(my_plan[[rn <- rn + 1, "target"]], "EfficiencyGraphs")
-  expect_equal(my_plan[[rn <- rn + 1, "target"]], "ExergyEnergyGraphs")
 })